refactor: separate logic into use-edit-board hook
