+++ src/modules/boards/edit-board/hooks/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-edit-board-loader";
+export * from "./use-edit-board";
